refactor(dev): move js and css compilers into assets namespace
