+++ packages/remix-dev/compiler/assets/js.ts
