--- packages/remix-dev/compiler/browserjs/index.ts
@@ -1,1 +1,0 @@
-export * as compiler from "./compiler";
