--- packages/remix-dev/compiler/css/index.ts
@@ -1,1 +1,0 @@
-export * as compiler from "./compiler";
